+++ app/src/main/java/com/eightman/kweather/utils/DateTimeUtils.kt
@@ -3,8 +3,8 @@ package com.eightman.kweather.utils
 import java.text.SimpleDateFormat
 import java.util.*
 
-fun Int.getUiDate(): String =
+fun Long.getUiDate(): String =
     SimpleDateFormat("MMM dd, yyyy", Locale.US).let { simpleDateFormat ->
         simpleDateFormat.timeZone = TimeZone.getTimeZone("GMT")
-        return simpleDateFormat.format(Date(this.toLong()))
+        return simpleDateFormat.format(Date(this))
     }
